Move ColumnMapping to single file
