+++ be/src/olap/schema_change.h
@@ -41,17 +41,6 @@ class RowCursor;
 // defined in 'writer.h'
 class ColumnDataWriter;
 
-struct ColumnMapping {
-    ColumnMapping() : ref_column(-1), default_value(NULL) {}
-    virtual ~ColumnMapping() {}
-
-    // <0: use default value
-    // >=0: use origin column
-    int32_t ref_column;
-    // normally for default value. stores values for filters
-    WrapperField* default_value;
-};
-
 class RowBlockChanger {
 public:
     typedef std::vector<ColumnMapping> SchemaMapping;
